+++ indic_ai/utils/models.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+class VoiceInput(BaseModel):
+    temple_name: str
+
+class BlogInput(BaseModel):
+    blog_url: str
+
+class ImageInput(BaseModel):
+    temple_name: str
